pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = ""
authors = [
    {name = "nananaseek",email = "nooobleemonk@gmail.com"}
]
# readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai (>=1.21.1,<2.0.0)",
    "instaloader (>=4.14.1,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiogram (==3.18)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "tiktokapi (>=7.1.0,<8.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
